+++ src/analogout.js
@@ -2,6 +2,9 @@ const { STATUS, STATUS_MSG } = require('./util/status');
 
 module.exports = RED => {
 	RED.nodes.registerType('Analog OUT', function (config) {
+		const topic = config.name || 'Analog OUT';
+		const portinfoType = '6';
+
 		RED.nodes.createNode(this, config);
 
 		let lastStatusString = '';
@@ -22,18 +25,23 @@ module.exports = RED => {
 		let isValidClamp = true;
 		const webio = RED.nodes.getNode(config.webio);
 		if (webio && webio.emitter) {
-			const topic = config.name || 'Analog OUT';
-			const portinfoType = '6';
 			let value;
 			let unit = '';
 			let clampLabels = [];
+			let clampData = null;
 
 			sendStatus(STATUS_MSG[STATUS.NOT_INITIALIZED]);
 
 			webio.emitter.addListener('webioLabels', labels => {
 				clampLabels = labels[portinfoType] || {};
 				isValidClamp = !!clampLabels[config.number];
-				this.send({ topic: topic, payload: value, unit: unit, clampName: clampLabels[config.number] || config.number });
+				const msg = { topic, payload: value, unit, clampName: clampLabels[config.number] || config.number };
+				Object.assign(msg, clampData);
+				this.send(msg);
+			});
+
+			webio.emitter.addListener('webioData', data => {
+				clampData = (data[portinfoType] || {})[config.number];
 			});
 
 			webio.emitter.addListener('webioGet', (type, values, status) => {
@@ -58,7 +66,9 @@ module.exports = RED => {
 
 					if (tmpValue !== value) {
 						value = tmpValue;
-						this.send({ topic: topic, payload: value, unit: unit, clampName: clampLabels[config.number] || config.number });
+						const msg = { topic, payload: value, unit, clampName: clampLabels[config.number] || config.number };
+						Object.assign(msg, clampData);
+						this.send(msg);
 					}
 				} else if (!isValidClamp && status === STATUS.OK) {
 					sendStatus(STATUS_MSG[STATUS.OK]);  // invalid clamp status (if invalid web-io configured)
@@ -68,6 +78,7 @@ module.exports = RED => {
 			this.on('close', () => {
 				webio.emitter.removeAllListeners('webioGet');
 				webio.emitter.removeAllListeners('webioLabels');
+				webio.emitter.removeAllListeners('webioData');
 				RED.comms.publish('wut/i18n-status/analogout/' + this.id, null, false); // publish empty message to "delete" retained message
 			});
 		} else {