+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Analog OUT node added
 - msg.clampName parameter for Web-IO output messages
+- display measurement value ranges for analog clamps (in config dialog + as parameters in output messages)
 
 ### Changed